+++ software_esp32/include/globals.h
@@ -42,6 +42,8 @@
 
 #include <Arduino.h>
 
+#define CHECK_BIT(var,pos) ((var>>pos) & 1)
+
 // #define CONFIG_ASYNC_TCP_USE_WDT 0
 
 #define ACTUATOR_COUNT      12