+++ software_esp32/src/web.cpp
@@ -107,10 +107,14 @@ String CWeb::getProtConfig (VDM_PROTOCOL_CONFIG protConfig)
                     "\"ip\":\""+ip2String(protConfig.brokerIp)+"\","+
                     "\"port\":\""+String(protConfig.brokerPort)+"\","+
                     "\"interval\":"+String(protConfig.brokerInterval)+","+
-                    "\"pubTarget\":"+String(protConfig.publishTarget)+","+
-                    "\"pubAllTemps\":"+String(protConfig.publishAllTemps)+","+
-                    "\"pubPathAsRoot\":"+String(protConfig.publishPathAsRoot)+","+
-                    "\"pubUpTime\":"+String(protConfig.publishUpTime)+","+
+                    "\"publish\":"+String(protConfig.publishInterval)+","+
+                    "\"pubMinDelay\":"+String(protConfig.minBrokerDelay)+","+
+                    "\"pubOnChange\":"+String(protConfig.protocolFlags.publishOnChange)+","+
+                    "\"pubTarget\":"+String(protConfig.protocolFlags.publishTarget)+","+
+                    "\"pubAllTemps\":"+String(protConfig.protocolFlags.publishAllTemps)+","+
+                    "\"pubPathAsRoot\":"+String(protConfig.protocolFlags.publishPathAsRoot)+","+
+                    "\"pubUpTime\":"+String(protConfig.protocolFlags.publishUpTime)+","+
+                    "\"keepAliveTime\":"+String(protConfig.keepAliveTime)+","+
                     "\"user\":\""+String(protConfig.userName)+"\"}";  
   return result;  
 }
@@ -301,10 +305,14 @@ String CWeb::getValvesStatus()
                  if (StmApp.actuators[x].temp2>-500) {
                     result +=",\"temp2\":" + String(((float)StmApp.actuators[x].temp2)/10,1);
                  }
-                 //if (Mqtt.mqttReceived || ServerServices.jsonSetValveReceived) {
                  if (controlActive) {
+                  if (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[x].link==0) {
                     result +=",\"tTarget\":" + String(((float)PiControl[x].target),1);
                     result +=",\"tValue\":" + String(((float)PiControl[x].value),1);
+                  } else {
+                    result +=",\"tTarget\":\"link #"+String(VdmConfig.configFlash.valvesControlConfig.valveControlConfig[x].link)+"\"";
+                    result +=",\"tValue\":\"link #"+String(VdmConfig.configFlash.valvesControlConfig.valveControlConfig[x].link)+"\"";
+                  }
                  }
                  result +="}";      
     start = true;